requirements.txt
streamlit
feedparser
requests
google-genai
python-dotenv
pandas
plotly
